Break out functionality to separate files
